requirements.txt
selenium==4.28.1
psycopg2-binary
pandas